working through scintilla

+++ StatTag/View Controllers/ScintillaHelpers/SCScintilla.h
@@ -20,8 +20,8 @@
 @property (strong, nonatomic) NSMutableArray* Markers;
 
 -(void)EmptyUndoBuffer;
--(void)LineScroll:(int)lines columns:(int)columns;
--(int)LinesOnScreen;
+-(void)LineScroll:(NSInteger)lines columns:(NSInteger)columns;
+-(NSInteger)LinesOnScreen;
 -(id)initWithScintillaView:(ScintillaView*)sc;
 
 -(SCLineCollection*)Lines;
